Backend_IA_mx32/requirements.txt
# Dependencias base
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
loguru==0.7.2
requests==2.32.3
aiohttp==3.9.5

# Firebase
firebase-admin==6.4.0
google-cloud-firestore==2.13.1

# LangChain y dependencias
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15
langchain-community==0.3.7

# Cerebras (opcional, si quieres usar el SDK oficial)
# cerebras-cloud-sdk

# Dependencias adicionales
python-multipart==0.0.9

